requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
requests>=2.26.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0 